feat: implement comment repo logic

+++ internal/domain/ports/accommodation.ports.go
@@ -6,7 +6,7 @@ type AccommodationOutput interface {
 	Create(data entities.Accommodation) (entities.Accommodation, *entities.Exception)
 	Update(id string, data entities.Accommodation) (entities.Accommodation, *entities.Exception)
 	Delete(id string) *entities.Exception
-	List(listing entities.Listing) ([]entities.Accommodation, int64, *entities.Exception)
+	List(listing entities.Listing) ([]entities.Accommodation, int, *entities.Exception)
 	FindOne(id string) (entities.Accommodation, *entities.Exception)
 }
 
@@ -14,6 +14,6 @@ type AccommodationInput interface {
 	Create(data entities.Accommodation) (entities.Accommodation, *entities.Exception)
 	Update(id string, data entities.Accommodation) (entities.Accommodation, *entities.Exception)
 	Delete(id string) *entities.Exception
-	List(listing entities.Listing) ([]entities.Accommodation, int64, *entities.Exception)
+	List(listing entities.Listing) ([]entities.Accommodation, int, *entities.Exception)
 	FindOne(id string) (entities.Accommodation, *entities.Exception)
 }
