+++ internal/domain/ports/authorizations.ports.go
@@ -6,7 +6,7 @@ type AuthorizationsOutput interface {
 	Create(data entities.Authorizations) (entities.Authorizations, *entities.Exception)
 	Update(id string, data entities.Authorizations) (entities.Authorizations, *entities.Exception)
 	Delete(id string) *entities.Exception
-	List() ([]entities.Authorizations, int64, *entities.Exception)
+	List() ([]entities.Authorizations, int, *entities.Exception)
 	FindOne(id string) (entities.Authorizations, *entities.Exception)
 }
 
@@ -14,6 +14,6 @@ type AuthorizationsInput interface {
 	Create(data entities.Authorizations) (entities.Authorizations, *entities.Exception)
 	Update(id string, data entities.Authorizations) (entities.Authorizations, *entities.Exception)
 	Delete(id string) *entities.Exception
-	List() ([]entities.Authorizations, int64, *entities.Exception)
+	List() ([]entities.Authorizations, int, *entities.Exception)
 	FindOne(id string) (entities.Authorizations, *entities.Exception)
 }
