+++ internal/domain/ports/booking.ports.go
@@ -6,7 +6,7 @@ type BookingOutput interface {
 	Create(data entities.Booking) (entities.Booking, *entities.Exception)
 	Update(id string, data entities.Booking) (entities.Booking, *entities.Exception)
 	Delete(id string) *entities.Exception
-	List(listing entities.Listing) ([]entities.Booking, int64, *entities.Exception)
+	List(listing entities.Listing) ([]entities.Booking, int, *entities.Exception)
 	FindOne(id string) (entities.Booking, *entities.Exception)
 }
 
@@ -14,6 +14,6 @@ type BookingInput interface {
 	Create(data entities.Booking) (entities.Booking, *entities.Exception)
 	Update(id string, data entities.Booking) (entities.Booking, *entities.Exception)
 	Delete(id string) *entities.Exception
-	List(listing entities.Listing) ([]entities.Booking, int64, *entities.Exception)
+	List(listing entities.Listing) ([]entities.Booking, int, *entities.Exception)
 	FindOne(id string) (entities.Booking, *entities.Exception)
 }
