+++ internal/domain/ports/comment.ports.go
@@ -6,7 +6,7 @@ type CommentOutput interface {
 	Create(data entities.Comment) (entities.Comment, *entities.Exception)
 	Update(id string, data entities.Comment) (entities.Comment, *entities.Exception)
 	Delete(id string) *entities.Exception
-	List(listing entities.Listing) ([]entities.Comment, int64, *entities.Exception)
+	List(listing entities.Listing) ([]entities.Comment, int, *entities.Exception)
 	FindOne(id string) (entities.Comment, *entities.Exception)
 }
 
@@ -14,6 +14,6 @@ type CommentInput interface {
 	Create(data entities.Comment) (entities.Comment, *entities.Exception)
 	Update(id string, data entities.Comment) (entities.Comment, *entities.Exception)
 	Delete(id string) *entities.Exception
-	List(listing entities.Listing) ([]entities.Comment, int64, *entities.Exception)
+	List(listing entities.Listing) ([]entities.Comment, int, *entities.Exception)
 	FindOne(id string) (entities.Comment, *entities.Exception)
 }
